Clean up unnecessarily files
