+++ src/scenes/Lines/scenes/Editor/JourneyPatterns/Editor/styles.scss
@@ -1,59 +1,4 @@
 @import '~@entur/tokens/dist/styles.scss';
-.journey-pattern-editor {
-  section {
-    h3 {
-      margin-bottom: 0;
-    }
-    p {
-      margin-top: 0;
-      svg {
-        color: $colors-validation-sky;
-      }
-    }
-  }
-}
-
-.journey-pattern-inputs {
-  display: flex;
-  flex-wrap: wrap;
-
-  .eds-input-group {
-    flex: 40% 1;
-    margin-right: 2rem;
-  }
-
-  .eds-input-group + .eds-input-group {
-    margin-top: 0;
-  }
-}
-
-.input-group {
-  margin-bottom: 4rem;
-
-  .input-fields {
-    display: flex;
-    flex-wrap: wrap;
-
-    div {
-      flex: 40%;
-      margin-right: 2rem;
-      margin-top: 0;
-    }
-  }
-
-  section {
-    padding: 0 3.75rem;
-    h3 {
-      margin-bottom: 0;
-    }
-    p {
-      margin-top: 0;
-      svg {
-        color: $colors-validation-sky;
-      }
-    }
-  }
-}
 
 .journey-pattern-inputs {
   display: flex;
@@ -87,4 +32,8 @@
 .operator-selector {
   max-width: 200px;
   width: 100%;
+}
+
+.stop-point-info {
+  margin-bottom: 2rem;
 }
